+++ foliant/preprocessors/dbdoc/oracle/main.py
@@ -6,7 +6,7 @@ from pkg_resources import resource_filename
 from foliant.preprocessors.utils.combined_options import CombinedOptions
 from logging import getLogger
 from .queries import (TablesQuery, ColumnsQuery, ForeignKeysQuery,
-                      FunctionsQuery, TriggersQuery)
+                      FunctionsQuery, TriggersQuery, ViewsQuery)
 
 logger = getLogger('flt.dbdoc.pgsql')
 
@@ -23,7 +23,8 @@ def process(config, tag_options) -> str:
         'components': [
             'tables',
             'functions',
-            'triggers'
+            'triggers',
+            'views'
         ]
     }
     options = CombinedOptions(
@@ -34,6 +35,8 @@ def process(config, tag_options) -> str:
         priority='tag',
         defaults=defaults
     )
+
+    logger.debug(f'Config: {options}')
 
     con = connect(options)
     return gen_docs(options, con)
@@ -107,20 +110,21 @@ def collect_datasets(connection,
         # fill each table with columns and foreign keys
         result['tables'] = collect_tables(tables, columns, fks)
 
-    # q_functions = FunctionsQuery(connection, filters)
-    # logger.debug(f'Functions keys query:\n\n {q_functions.sql}')
-    # functions = q_functions.run()
+    if 'views' in components:
+        q_views = ViewsQuery(connection, filters)
+        logger.debug(f'Views query:\n\n {q_views.sql}')
+        result['views'] = q_views.run()
 
-    # q_parameters = ParametersQuery(connection, filters)
-    # logger.debug(f'Parameters keys query:\n\n {q_parameters.sql}')
-    # parameters = q_parameters.run()
-
-    # # fill each function with its parameters
     if 'functions' in components:
-        result['functions'] = FunctionsQuery(connection, filters).run()
+        q_functions = FunctionsQuery(connection, filters)
+        logger.debug(f'Functions query:\n\n {q_functions.sql}')
+        result['functions'] = q_functions.run()
 
     if 'triggers' in components:
-        result['triggers'] = TriggersQuery(connection, filters).run()
+        q_triggers = TriggersQuery(connection, filters)
+        logger.debug(f'Triggers query:\n\n {q_triggers.sql}')
+        result['triggers'] = q_triggers.run()
+
     return result
 
 
@@ -178,11 +182,7 @@ def to_md(data: dict, template: str) -> str:
     with open(template, encoding='utf8') as f:
         template = Template(f.read())
 
-    return template.render(
-        tables=data['tables'],
-        functions=data['functions'],
-        triggers=data['triggers']
-    )
+    return template.render(**data)
 
 
 def to_diag(data: dict, template: str) -> str: