+++ changelog.md
@@ -1,3 +1,9 @@
+# 0.1.1
+
+-    New: Add views query to components
+-    Fix: Oracle triggers query
+-    Fix: Fix both PostgreSQL and Oracle templates
+
 # 0.1.0
 
 -    Initial release